Initial interface-neutral refactoring moving core CRDS modules into crds.core subpackage that is also spliced into the original crds namespace using sys.modules.
