+++ crds/core/utils.py
@@ -16,10 +16,11 @@ import datetime
 import ast
 import gc
 
-# from crds import data_file,  import deferred until required
+# from crds.core import data_file,  import deferred until required
 
-from crds import log, config, pysh, ALL_OBSERVATORIES, INSTRUMENT_KEYWORDS
-from crds.python23 import *
+from crds.core import log, config, pysh
+from crds.core.constants import ALL_OBSERVATORIES, INSTRUMENT_KEYWORDS
+from crds.core.python23 import *
 
 # ===================================================================
 
@@ -923,7 +924,7 @@ def file_to_instrument(filename):
     for (_obs, instr) in observatory_instrument_tuples():
         if ("_" + instr + "_" in basename) or basename.startswith(instr + "_"):
             return instr.upper()
-    from crds import data_file
+    from crds.core import data_file
     header = data_file.get_unconditioned_header(filename, needed_keys=INSTRUMENT_KEYWORDS)
     return header_to_instrument(header)
     
@@ -1027,7 +1028,7 @@ reference_to_locator = file_to_locator
 def test():
     """Run doctests."""
     import doctest
-    from crds import utils
+    from crds.core import utils
     return doctest.testmod(utils)
 
 if __name__ == "__main__":
