+++ setup.py
@@ -10,6 +10,7 @@ setup_pars = {
         'crds.bestrefs',
         'crds.certify',
         'crds.client',
+        'crds.core',
         'crds.hst',
         'crds.jwst',
         'crds.submit',
@@ -21,6 +22,7 @@ setup_pars = {
         'crds.bestrefs' : 'crds/bestrefs',
         'crds.certify' : 'crds/certify',
         'crds.client' : 'crds/client',
+        'crds.core' : 'crds/core',
         'crds.hst' : 'crds/hst',
         'crds.jwst' : 'crds/jwst',
         'crds.submit' : 'crds/submit',
